requirements.txt
requests
pandas
python-telegram-bot
apscheduler
pytz
websocket-client
numpy
base58